cleanup icons, modal fixing
