+++ src/App.jsx
@@ -1,15 +1,11 @@
 import React from "react";
-import Page from "./components/mainPage";
+
 import Desktop from "./components/desktop";
 import "@react95/core/GlobalStyle";
 import "@react95/core/themes/win95.css";
 
 const App = () => {
-  return (
-    <Page>
-      <Desktop />
-    </Page>
-  );
+  return <Desktop />;
 };
 
 export default App;
