--- src/components/desktopBox.jsx
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-const Box = styled.div({
-  alignItems: "flex-start",
-  display: "flex",
-  justifyContent: "center",
-  flexDirection: "column",
-});
-
-export default Box;
