--- src/components/iconBox.jsx
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-import Box from "./desktopBox";
-
-const IconBox = styled(Box)({
-  alignItems: "center",
-  justifyContent: "center",
-  marginBottom: 20,
-  padding: 8,
-});
-
-export default IconBox;
