--- src/components/socialMedia.jsx
@@ -1,20 +1,0 @@
-const socialMedia = [
-  {
-    name: "LinkedIn",
-    url: "https://www.linkedin.com/in/janneilkka/",
-  },
-  {
-    name: "Github",
-    url: "https://www.github.com/janneilkka/",
-  },
-  {
-    name: "Instagram",
-    url: "https://instagram.com/janne.ilkka",
-  },
-  {
-    name: "Spotify",
-    url: "https://open.spotify.com/user/rogergregor",
-  },
-];
-
-export default socialMedia;
