+++ src/components/useModal.jsx
@@ -5,12 +5,10 @@ const useModal = (initialState = false) => {
 
   const openModal = useCallback(() => {
     setIsOpen(true);
-    console.log("Modal true");
   }, []);
 
   const closeModal = useCallback(() => {
     setIsOpen(false);
-    console.log("Modal false");
   }, []);
 
   return [isOpen, openModal, closeModal];
